+++ src/Control/Monad/Time.hs
@@ -7,23 +7,18 @@
 module Control.Monad.Time (MonadTime(..)) where
 
 import Control.Monad.Trans
-import Control.Monad.Reader (ReaderT, ask)
 import Data.Time
+import GHC.Clock (getMonotonicTime)
 
--- | Class of monads which carry the notion of the current time.
+-- | Class of monads which make it possible to measure time.
 class Monad m => MonadTime m where
   currentTime :: m UTCTime
+  monotonicTime :: m Double
 
 -- | Base instance for IO.
 instance MonadTime IO where
   currentTime = getCurrentTime
-
--- | This is @ReaderT UTCTime@ on purpose, to avoid breaking
--- downstream.
---
--- @since 0.3.0.0
-instance {-# OVERLAPPING #-} Monad m => MonadTime (ReaderT UTCTime m) where
-  currentTime = ask
+  monotonicTime = getMonotonicTime
 
 -- | Generic, overlapping instance.
 instance {-# OVERLAPPABLE #-} (
@@ -32,3 +27,4 @@ instance {-# OVERLAPPABLE #-} (
   , Monad (t m)
   ) => MonadTime (t m) where
     currentTime = lift currentTime
+    monotonicTime = lift monotonicTime