+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# monad-time-0.4.0.0 (2022-07-05)
+* Added `monotonicTime` to `MonadTime`
+* Removed the `MonadTime` instance for `ReaderT UTCTime`
+
 # monad-time-0.3.1.0 (2018-04-09)
 * Restored compatibility with GHC versions < 7.10
   ([#4](https://github.com/scrive/monad-time/issues/4)).